move live inspect test to integration test file
